real_estate/requirements.txt
fastapi
uvicorn
requests
python-dotenv
agno
pydantic
firecrawl-py
openai
